refactor(typings): add and update all other payloads
